+++ main_t.c
@@ -1,5 +1,5 @@
 #include <stdio.h>
-#include "src/routine.h"
+#include "src/routine_user_api.h"
 
 int add(int a, int b);
 int routine_b(int k)
@@ -15,20 +15,36 @@ int routine_b(int k)
 int routine_add(int a, int b)
 {
     int c = a + b;
-    exchange();
     return c;
 }
 
 int main()
 {
-    create_current_routine();
-    create_current_routine();
     rid_t e = create_routine_with_params(routine_add, 2, 3, 5);
     rid_t r = create_routine_with_params(routine_b, 1, 10);
-    create_routine_with_params(routine_b, 1, 10);
+    pid_t k = create_routine_with_params(routine_b, 1, 10);
     exchange();
     exchange();
     exchange();
+    exchange();
+    block(r);
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    exchange();
+    resume(r);
     exchange();
     exchange();
     exchange();
