+++ src/routine_helper.c
@@ -1,47 +1,38 @@
-#include "routine.h"
+#include "routine_helper.h"
 
-void remove(proutine *h, pcollect *c,uroutine** u) {
-    pcollect t = malloc(sizeof(collect));
-    uroutine* u1 = malloc(sizeof(uroutine));
-    t->bs = (*h)->bs;
-
-    if ((*c) != NULL) {
-        t->link.next = (pcollect) (*c);
-    }
-    (*c) = t;
-
-    u1->consequence = (*h)->rax;
-    u1->status = T;
-    u[u1->rid=(*h)->rid]=u1;
-    remove_from_bitmap((*h)->rid);
-
-    proutine lh = (*h);
-    (*h) = (*h)->next;
-    free(lh);
-}
 
 void exchange_c(proutine *h, proutine *t) {
-    proutine n = (*h)->next;
-    if (n == NULL) {
-        return;
+    proutine n = *h;
+    proutine f = NULL, e = NULL;
+    for (;((n = n->next) == NULL) || (n->status == UB);) {
+        if (n == NULL) {
+            return;
+        }
+        f = f == NULL ? n : f;
+        e = n;
+    }
+    if(e!=NULL) {
+        e->next = NULL;
     }
     n->status = R;
     proutine curr = (*h);
-    curr->next = NULL;
+    curr->next = f;
     (*t)->next = curr;
-    (*t) = curr;
+    (*t) = e == NULL ? curr : e;
     (*h) = n;
 }
 
-void init_stack(proutine r, data_p stack, int len, any p) {
+void init_stack(proutine r, data_p stack, int len, any p, any stop) {
     r->bs.stack = stack;
     r->bs.size = len;
-    stack[len - 1] = (code_t) stop_routine;
+    stack[len - 1] = (code_t) stop;
     stack[len - 2] = (code_t) p;
-    r->esp = (code_t) (stack + STACK_LEN - 2);
-    r->ebp = (code_t) (stack + STACK_LEN - 1);
+    r->esp = (code_t) (stack + len - 2);
+    r->ebp = (code_t) (stack + len - 1);
 }
 
-void insert(proutine *n, proutine r) {
-    (*n) = (*n)->next = r;
+proutine init_routine() {
+    proutine r = malloc(sizeof(routine));
+    memset(r, 0, sizeof(routine));
+    return r;
 }
